Some little warning fixes and other adjustments

+++ LunaDll/GlobalFuncs.cpp
@@ -212,7 +212,7 @@ std::wstring GetNonANSICharsFromWStr(const std::wstring& wstr)
     std::wstring ansi = StrA2WStr(WStr2StrA(wstr));
     std::wstring mismatches;
 
-    for (int i = 0, j = 0; (i < wstr.size()) && (j < ansi.size()); i++)
+    for (size_t i = 0, j = 0; (i < wstr.size()) && (j < ansi.size()); i++)
     {
         if (wstr[i] != ansi[j])
         {
